Added comment-groups.csv export.
Colin will be using this data (or some filtered version of it) to pass to an
LLM when it wants to summarize things.

The code uses the summarized data from the PCA json blob instead of computing
things from the raw comments and votes tables. The latter approach results in
numbers that don't match up exactly with the data that appears on the HTML
version of the report (our numbers are a little higher, so the Clojure backend
is filtering out some votes/voters that we are not).

We want the LLM to see the exact same data that's on the HTML page because it
might refer to specific numbers and we want those numbers to be exactly the
same as the numbers the user sees.