Fix dev environment startup (#2211)
* Revert "Bump torch from 2.3.1 to 2.8.0 in /delphi (#2142)"

This reverts commit a7a060b8b63372141a6b092848d4823b8d8b9c0d.

* Move clojure math env to dev instead of prod

Set `MATH_ENV=dev` instead of `=prod` in `example.env`. This avoids
an infinite reboot loop of the clojure worker due to failing to load Datadog
profiler -- which is skipped in development environment.

* Start notes to get running

@ballPointPenguin has asked me to let him know whether
`make start` works as intended. Documenting here the steps needed
to make it work :)

* Describe fix for login problem

* Fix login failure due to missing hostname in certificate

* Remove explanatory notes to make a clean commit

As discussed with @ballPointPenguin

---------

Co-authored-by: Julien Cornebise <julien@cornebise.com>